+++ main.py
@@ -1,17 +1,10 @@
+from utils.model import research_stations
+from utils.controller import get_stations_info
 
 
-research_stations : list =[
-
-    {"name" : "Instytut_Centrum_Zdrowia_Matki_Polki" , "location": "Łódź" },
-    {"name" : "Instytut_Fizjologii_i_Patologii_Słuchu" , "location": "Warszawa"},
-    {"name" : "Instytut_Chemicznej_Przeróbki_Węgla" , "location": "Zabrze"},
-    {"name" : "Instytut_Wymiaru_Sprawiedliwości" , "location": "Warszawa" },
-
-]
-
-def get_stations_info( stations_data : list ) -> None:
-    for station in stations_data:
-        print(f"{station['name']} : {station['location']}" )
+def main():
+    get_stations_info(research_stations)
 
 
-get_stations_info( research_stations )
+if __name__ == "__main__":
+    main()
