+++ utils/model.py
@@ -1,0 +1,8 @@
+research_stations : list =[
+
+    {"name" : "Instytut_Centrum_Zdrowia_Matki_Polki" , "location": "Łódź" },
+    {"name" : "Instytut_Fizjologii_i_Patologii_Słuchu" , "location": "Warszawa"},
+    {"name" : "Instytut_Chemicznej_Przeróbki_Węgla" , "location": "Zabrze"},
+    {"name" : "Instytut_Wymiaru_Sprawiedliwości" , "location": "Warszawa" },
+
+]
